requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
alembic==1.6.5
amqp==5.0.6
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
cchardet==2.1.7
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
dnspython==2.1.0
elasticsearch==7.13.0
email-validator==1.1.2
fake-useragent==0.1.11
flashtext==2.7
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gender-guesser==0.4.0
geographiclib==1.50
geopy==2.1.0
googletransx==2.4.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.3.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==2.11.3
joblib==1.0.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
multidict==5.1.0
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pika==0.13.0
pkg-resources==0.0.0
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pycares==4.0.0
pycountry==20.7.3
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-editor==1.0.4
python-socks==1.2.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
schedule==1.1.0
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
threadpoolctl==2.1.0
toml==0.10.2
torpy==1.1.6
tqdm==4.61.0
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
yarl==1.6.3
zipp==3.4.1
